+++ packages/example/src/components/pages/Theme.js
@@ -9,9 +9,9 @@ class Theme extends Component {
   render() {
     return (
       <div>
-        <Field type="text" name="username" label="Username*" value="" className={styles.theme} required />
-        <Field type="text" name="test-field" label="Testing" value="" className={styles.theme} />
-        <Field type="text" name="test-field2" label="Testing" value="" className={styles.underline} size={100} />
+        <Field type="text" name="username" label="Username*" defaultValue="" className={styles.theme} required />
+        <Field type="text" name="test-field" label="Testing" defaultValue="" className={styles.theme} />
+        <Field type="text" name="test-field2" label="Testing" defaultValue="" className={styles.underline} size={100} />
         <Radios
           name="radio-testing"
           type="radios"
@@ -36,12 +36,12 @@ class Theme extends Component {
           defaultValue="second"
         />
 
-        <Field name="checkbox1" type="checkbox" value="something" className={styles.theme} label="Checkbox 1" />
-        <Field name="checkbox2" type="checkbox" value="something" className={styles.theme} label="Checkbox 2" />
+        <Field name="checkbox1" type="checkbox" defaultValue="something" className={styles.theme} label="Checkbox 1" />
+        <Field name="checkbox2" type="checkbox" defaultValue="something" className={styles.theme} label="Checkbox 2" />
         <Field
           name="checkbox3"
           type="checkbox"
-          value="something"
+          defaultValue="something"
           className={styles.theme}
           label="Checkbox 3"
           disabled
@@ -50,21 +50,21 @@ class Theme extends Component {
         <Field
           name="checkbox4"
           type="checkbox"
-          value="something"
+          defaultValue="something"
           label="Checkbox 4"
           className={classes(styles.theme, styles.inline)}
         />
         <Field
           name="checkbox5"
           type="checkbox"
-          value="something"
+          defaultValue="something"
           label="Checkbox 5"
           className={classes(styles.theme, styles.inline)}
         />
         <Field
           name="checkbox6"
           type="checkbox"
-          value="something"
+          defaultValue="something"
           label="Checkbox 6"
           className={classes(styles.theme, styles.inline)}
           disabled
@@ -72,42 +72,20 @@ class Theme extends Component {
         <Field
           name="checkbox7"
           type="checkbox"
-          value="something"
+          defaultValue="something"
           label="Checkbox 7"
           className={classes(styles.theme, styles.inline)}
           disabled
           defaultChecked
         />
 
-        <Field
-          name="switch1"
-          type="checkbox"
-          value="something"
-          label="Switch 1"
-          className={classes(styles.switch)}
-          defaultChecked
-        />
+        <Field name="switch1" type="checkbox" label="Switch 1" className={styles.switch} defaultChecked />
 
-        <Field name="switch2" type="checkbox" value="something" label="Switch 2" className={classes(styles.switch)} />
+        <Field name="switch2" type="checkbox" label="Switch 2" className={styles.switch} />
 
-        <Field
-          name="switch3"
-          type="checkbox"
-          value="something"
-          label="Switch 3"
-          className={classes(styles.switch)}
-          disabled
-          defaultChecked
-        />
+        <Field name="switch3" type="checkbox" label="Switch 3" className={styles.switch} disabled defaultChecked />
 
-        <Field
-          name="switch4"
-          type="checkbox"
-          value="something"
-          label="Switch 4"
-          className={classes(styles.switch)}
-          disabled
-        />
+        <Field name="switch4" type="checkbox" label="Switch 4" className={styles.switch} disabled />
       </div>
     );
   }