+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.5.0
+
+- [helpers] add `getCursor` and `hasSelectionStart`
+- [field] by passing a `value` prop, you can now control a field
+- [all] upgrade devDeps
+- [all] convert to react-testing-library instead of enzyeme
+
 ## 0.4.3
 
 - [field] fix attempt to unregister when not registered